Moved test from pkg to spectrocloud to run unit test
